dataviewer: disable (broken) edit
The edit functionality was previously broken due to
encoding issues, so we just replace the column type
with a generic read-only variant.
